requirements.txt
Flask
gunicorn
psycopg2
djangorestframework-jwt==1.11.0